Update code to use pdb and pdbRedo databases
The code is modified to fetch and display data from both 'pdb' and 'pdbRedo' databases. To achieve this, new variables are defined to store the results from each database and fetch operations are carried out separately for each URL. This allows for a more robust display of data in cases where an entry is not found in one database but exists in the other. Also standardize the accessor names to camelCase in 'Constants.tsx' file.