Add statistics page with glycans graph representation
The commit introduces a new page for displaying various statistics related to glycans. It also adds a new plotly graph showing Glycosylation in the PDB over time. Modifications have been made to the navigation bar to link to this newly created statistics page and a new file for the statistics header was added. Code for fetching and processing the data for the graph has also been implemented.