Update error graphs and add image export option
Enhancements were made in the error graphs to include a dynamic range for years and an export option was introduced allowing users to save the graphs as an SVG file. The export configuration includes settings for filename, dimensions, and format. This new functionality enables better data analysis and flexibility in terms of usage and distribution of the generated graphs.

Signed-off-by: Jordan Dialpuri <44945647+Dialpuri@users.noreply.github.com>