Modify graph background and add modebar styling
The background color of the error graphs was changed to transparent from preset colors and a modebar was added with defined colors. Additionally, minor adjustments were done in some graph elements like adding a suffix in yaxis2, adjusting the range of xaxis and fixing code styling issues. These changes improve the visual representation of data as well as the code readability.

Signed-off-by: Jordan Dialpuri <44945647+Dialpuri@users.noreply.github.com>