Refactor CMakeLists and remove unnecessary code blocks
In this commit, the arrangement of the files in CMakeLists.txt has been significantly simplified. In gemmi/CMakeLists.txt, the static library gemmi_cpp is changed to a regular library and globbing is used to include header files instead of listing them individually. Similarly, in the main CMakeLists.txt, unnecessary code blocks for testing mode or ccp4srs have been removed and some modifications have been done on setting flags and file paths.